fix bug in Player constructor touch up DBATest.java

+++ test/DBATest.java
@@ -37,15 +37,6 @@ public class DBATest {
             new Player(400, 2,
                     "B", 0, false, "INJAILROUND1"),
             new Player(1000, 2,
-                    "C", 0, false, "FREE")
-    };
-
-    static Player[] players3 = {
-            new Player(100, 2,
-                    "A", 0, false, "FREE"),
-            new Player(400, 2,
-                    "B", 0, false, "INJAILROUND1"),
-            new Player(1000, 2,
                     "C", 0, false, "FREE"),
             new Player(-200, 2,
                     "D", 0, true, "FREE"),
@@ -61,37 +52,57 @@ public class DBATest {
         // delete existing saved game data
         try {
             File file = new File(DBAccessor.DB_FILE_PATH);
-            if (file.delete()) {
-                System.out.println(file.getName() + " deleted successfully.");
-            } else {
-                System.out.println("Delete file failed.");
+            if (file.exists()) {
+                if (file.delete()) {
+                    System.out.println(file.getName() + " deleted successfully.");
+                } else {
+                    System.out.println("Delete file failed.");
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
         }
 
+        // test 1
+        // create a game
         GlobalController.gameController = new GameController(players1);
+        // write down current static game values
+        int round = GameController.round;
+        int whosTurn = GameController.whosTurn;
+        Player[] players = GameController.players.clone();
+        // save game
         DBAccessor.SaveGame();
-        GameController loaded_gc = DBAccessor.LoadGame();
-        assert GlobalController.gameController.equals(loaded_gc);
-        // save the games
-//        DBAccessor.SaveGame(historyGame1);
-//        DBAccessor.SaveGame(historyGame2);
-//        DBAccessor.SaveGame(historyGame3);
+        // load saved game
+        GlobalController.gameController = DBAccessor.LoadGame();
+        // run test
+        assert GameController.round==round;
+        assert GameController.whosTurn==whosTurn;
+        assert players.length == GameController.players.length;
+        for (int i=0;i<players.length;++i){
+            assert players[i].equals(GameController.players[i]);
+        }
 
-        // load the history game list. The getHistoryGameList will return a lighted historyGames class to tell the user
-        // basic information like stored data, name, player names etc.
-//        HistoryGame[] historyGames = DBAccessor.getHistoryGameList();
-//        assert historyGames[0].gameListToString().equals(historyGame1.gameListToString());
-//        assert historyGames[1].gameListToString().equals(historyGame2.gameListToString());
-//        assert historyGames[2].gameListToString().equals(historyGame3.gameListToString());
+        // test 2
+        // create a game
+        GlobalController.gameController = new GameController(players2);
+        // modify game states
+        GameController.whosTurn = 3;
+        GameController.round = 99;
+        // write down current static game values
+        round = GameController.round;
+        whosTurn = GameController.whosTurn;
+        players = GameController.players.clone();
+        // save game
+        DBAccessor.SaveGame();
+        // load saved game
+        GlobalController.gameController = DBAccessor.LoadGame();
+        // run test
+        assert GameController.round==round;
+        assert GameController.whosTurn==whosTurn;
+        assert players.length == GameController.players.length;
+        for (int i=0;i<players.length;++i){
+            assert players[i].equals(GameController.players[i]);
+        }
 
-        // gameListToString and toString will well format all the information including player information and square
-        // infomation, so that only if these 2 are "equal", the test will past.
-
-        // load and compare the history games.
-//        assert DBAccessor.LoadGame(1).toString().equals(historyGame1.toString());
-//        assert DBAccessor.LoadGame(2).toString().equals(historyGame2.toString());
-//        assert DBAccessor.LoadGame(3).toString().equals(historyGame3.toString());
     }
 }
